start trying to move utils into utils
