+++ darkside-tests/src/constants.rs
@@ -1,4 +1,4 @@
-use crate::darkside::darkside_types::TreeState;
+use crate::darkside_types::TreeState;
 
 pub const GENESIS_BLOCK: &'static str = "0400000027e30134d620e9fe61f719938320bab63e7e72c91b5e23025676f90ed8119f02c94c1ef3fa94b348f1b6d1bdf5899ad1eac7bfbf0de6d2cd1a4bedf84699b09c415071fecf148b458b159d1d310b2d072e8250bfefcd1a7ead2b309071f6e954ecf581640f0f0f201b0092b34257e0a52315e66f950db025ae0b0b3b6c71aa017c8a2348b4dd00002400f5c2c1b120d8abc30adf542ee48f6301eb0f730839c15bfcdbb42df0a8d4a326cd9389010400008085202f89010000000000000000000000000000000000000000000000000000000000000000ffffffff03510101ffffffff000000000001000000c041bfdaffffffff0001fc0ffba31852cdff6ccbbb61157848f9fb42e6a6eb831806a5e1955f21dc943c5fcb897f9314fc5ebac4510b86706f45458c6db1d36dbab6ee23315e9654c54cfdd8aaf1af233dee596a2f03d42315fde064eade886a2b10843807d9498e9a83604227bb3dcc516b05c1f87ff343099d89d8fe27196be88ea0f6a59503bbdf4992a740469cd449dd253e18eaeb6392e435e22d3d1dc75530b30a06f1ae3ea8dab48e637cea7747a571f7539e338461fd5938581878f9b6111b59de5c25d93043b99c8472d04bb0300eaa68eeba6ccac5de0d36fe25bf3639d58fbc2480b8fcab31dada9929d9bf44515c5d1517817047d22badd35a37f174b3803fe6a9dece7f11c93af94c33eb8b442bc94d728452a545ce35a41ad2580c8c737a245d5cdc5f36df5bdd4ed27ffb6568893c909c690c95d6db22f09680f4febe906d456070dc9363b497c67b854125572a811a24a65b63e191e09026b7409b71e18e5fe62d3d1840a557d4b222e07e00dac01131f0993a91914e70ed8f6a609f0c894f6eeaf3b35fcc830159851fff0d38ed3c0f15acb3d4de404ead0bfdd9c10d169cdd4bb1cf4a074349c330b57636b093a9b3b749ed14ef347326b313267f34be94ec9b9b7ae41d4ccd7364ed647a5a74b432cd3c063d5b2bbaedddba11e25107ec4e89c41ff45173ae49e3a323280aa9a3cdc8d9a7c264d1f08cc2bcc486d02fa2158054aaa55e82448e827a7a50778d04308a99401eeb7c835568c15ca2104bd05ebe569eed9ac64ef63143eeb4bf203ac6215cd63f921bfc2b3d704533cb32a82fbbe4caa677ee1a970e10dd4a36b4a8cd3003adc2962776a74aadb5577c457664a4970fa0c78d4edcb48e5743433d5f4d329148bf2d4a6c3f536ebad618ece61ebe09c60a8ec0ac73d9c6ba0429ac4a63d6b37db740fff665c9a2daf2b8c24f0269a9fe34ea99578926efa81b40da08d66bc4c04ccbf0f6a96a0bbda3966282af280600a686365ab3eb47a46a4204d40340aea39c0ebb910a786e0b88a115d9d9d95e1661d4548ba1cb5941082d3249aa37444b844fb1932e81d4427efad6ebc4addff3daf3246a0f77273aa1ad352fcbddf4fa617bcee201eae7c763737f2168d236fa764e17addf1ecb42706e30918378b1b2a0aca3233a129f55685264c6f5a49867fdaa73797616c785d18a707805c2514e91af9015f0cbf2097be74ee4ee81eed4723afd0d45f656c539e2e6e4578ad342d84a2d812199486e0d57094125551fd3a268a383c1f9642809858b23aa11a7e242012f49c0d66417177d4cd0de2ca4430929a7e50e9cd8b569cfd17cf648473e80cea7003e4aa0cc382d6bf962bed6a00cade090c0b43ff689dcd535085e2aac6fe807e6263b2564b86c469baeb3586c1e0db5039ced9ac723f3911e28e396655e825101";
 pub const DARKSIDE_SEED: &'static str =
